More fixes

+++ app/src/main/java/com/ferrariofilippo/saveapp/view/ManageTagsFragment.kt
@@ -3,6 +3,7 @@
 
 package com.ferrariofilippo.saveapp.view
 
+import android.app.Activity
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
@@ -124,38 +125,49 @@ class ManageTagsFragment : Fragment() {
 
     private fun onRemoveTagInvoked(item: Tag, position: Int) {
         val app = requireActivity().application as SaveAppApplication
+        val activity = requireActivity()
         if (item.id == INCOME_ID) {
+            (binding.tagsRecyclerView.adapter as TagsAdapter).notifyItemChanged(position)
+            showSnackbar(activity, R.string.cannot_delete_tag)
+
             return
         }
 
-        val activity = requireActivity()
         lifecycleScope.launch {
             val firstMatchingTag = app.transactionRepository.getFirstWithTag(item.id)
-
             if (firstMatchingTag != null) {
                 (binding.tagsRecyclerView.adapter as TagsAdapter).notifyItemChanged(position)
-
-                Snackbar.make(
-                    activity.findViewById(R.id.containerView),
-                    R.string.cannot_delete_tag,
-                    Snackbar.LENGTH_SHORT
-                ).setAnchorView(activity.findViewById(R.id.bottomAppBar)).show()
+                showSnackbar(activity, R.string.cannot_delete_tag_data)
             } else {
-                app.tagRepository.delete(item)
-                TagUtil.incomeTagIds.remove(item.id)
+                val firstChildren = app.tagRepository.getFirstChildren(item.id)
+                if (firstChildren != null) {
+                    (binding.tagsRecyclerView.adapter as TagsAdapter).notifyItemChanged(position)
+                    showSnackbar(activity, R.string.cannot_delete_tag_children)
+                } else {
+                    app.tagRepository.delete(item)
+                    TagUtil.incomeTagIds.remove(item.id)
 
-                Snackbar.make(
-                    activity.findViewById(R.id.containerView),
-                    R.string.tag_deleted,
-                    Snackbar.LENGTH_SHORT
-                )
-                    .setAction(R.string.undo) {
-                        lifecycleScope.launch {
-                            app.tagRepository.insert(item)
-                            TagUtil.incomeTagIds.add(item.id)
-                        }
-                    }.setAnchorView(activity.findViewById(R.id.bottomAppBar)).show()
+                    Snackbar.make(
+                        activity.findViewById(R.id.containerView),
+                        R.string.tag_deleted,
+                        Snackbar.LENGTH_SHORT
+                    )
+                        .setAction(R.string.undo) {
+                            lifecycleScope.launch {
+                                app.tagRepository.insert(item)
+                                TagUtil.incomeTagIds.add(item.id)
+                            }
+                        }.setAnchorView(activity.findViewById(R.id.bottomAppBar)).show()
+                }
             }
         }
+    }
+
+    private fun showSnackbar(act: Activity, textRes: Int) {
+        Snackbar.make(
+            act.findViewById(R.id.containerView),
+            textRes,
+            Snackbar.LENGTH_SHORT
+        ).setAnchorView(act.findViewById(R.id.bottomAppBar)).show()
     }
 }
